chore: move room exception package
