+++ src/main/kotlin/smalltalk/backend/infra/repository/room/RedisRoomRepository.kt
@@ -6,9 +6,9 @@ import org.springframework.data.redis.connection.RedisConnection
 import org.springframework.data.redis.core.StringRedisTemplate
 import org.springframework.stereotype.Repository
 import smalltalk.backend.domain.room.Room
-import smalltalk.backend.presentation.exception.room.situation.FullRoomException
-import smalltalk.backend.presentation.exception.room.situation.RoomIdNotGeneratedException
-import smalltalk.backend.presentation.exception.room.situation.RoomNotFoundException
+import smalltalk.backend.exception.room.situation.FullRoomException
+import smalltalk.backend.exception.room.situation.RoomIdNotGeneratedException
+import smalltalk.backend.exception.room.situation.RoomNotFoundException
 
 
 @Repository
